aiApi/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pandas>=2.2.2
scikit-learn>=1.4.2
numpy>=1.26.4
python-multipart>=0.0.9
mysql-connector-python>=8.3.0